+++ apps/sim/tools/salesforce/utils.ts
@@ -1,0 +1,70 @@
+import { createLogger } from '@/lib/logs/console/logger'
+
+const logger = createLogger('SalesforceUtils')
+
+/**
+ * Extracts Salesforce instance URL from ID token or uses provided instance URL
+ * @param idToken - The Salesforce ID token containing instance URL
+ * @param instanceUrl - Direct instance URL if provided
+ * @returns The Salesforce instance URL
+ * @throws Error if instance URL cannot be determined
+ */
+export function getInstanceUrl(idToken?: string, instanceUrl?: string): string {
+  if (instanceUrl) return instanceUrl
+  if (idToken) {
+    try {
+      const base64Url = idToken.split('.')[1]
+      const base64 = base64Url.replace(/-/g, '+').replace(/_/g, '/')
+      const jsonPayload = decodeURIComponent(
+        atob(base64)
+          .split('')
+          .map((c) => `%${(`00${c.charCodeAt(0).toString(16)}`).slice(-2)}`)
+          .join('')
+      )
+      const decoded = JSON.parse(jsonPayload)
+      if (decoded.profile) {
+        const match = decoded.profile.match(/^(https:\/\/[^/]+)/)
+        if (match) return match[1]
+      } else if (decoded.sub) {
+        const match = decoded.sub.match(/^(https:\/\/[^/]+)/)
+        if (match && match[1] !== 'https://login.salesforce.com') return match[1]
+      }
+    } catch (error) {
+      logger.error('Failed to decode Salesforce idToken', { error })
+    }
+  }
+  throw new Error('Salesforce instance URL is required but not provided')
+}
+
+/**
+ * Extracts a descriptive error message from Salesforce API responses
+ * @param data - The response data from Salesforce API
+ * @param status - HTTP status code
+ * @param defaultMessage - Default message to use if no specific error found
+ * @returns Formatted error message
+ */
+export function extractErrorMessage(data: any, status: number, defaultMessage: string): string {
+  if (Array.isArray(data) && data[0]?.message) {
+    return `Salesforce API Error (${status}): ${data[0].message}${data[0].errorCode ? ` [${data[0].errorCode}]` : ''}`
+  }
+  if (data?.message) {
+    return `Salesforce API Error (${status}): ${data.message}`
+  }
+  if (data?.error) {
+    return `Salesforce API Error (${status}): ${data.error}${data.error_description ? ` - ${data.error_description}` : ''}`
+  }
+  switch (status) {
+    case 400:
+      return `Salesforce API Error (400): Bad Request - The request was malformed or missing required parameters`
+    case 401:
+      return `Salesforce API Error (401): Unauthorized - Invalid or expired access token. Please re-authenticate.`
+    case 403:
+      return `Salesforce API Error (403): Forbidden - You do not have permission to access this resource.`
+    case 404:
+      return `Salesforce API Error (404): Not Found - The requested resource does not exist or you do not have access to it.`
+    case 500:
+      return `Salesforce API Error (500): Internal Server Error - An error occurred on Salesforce's servers.`
+    default:
+      return `${defaultMessage} (HTTP ${status})`
+  }
+}
